Added navigationBar and navigationPopover comps
